requirements.txt
# EVA - Vehicle Detection & Tracking Evaluation
# Python >= 3.8

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
numpy>=1.24.0

# Tracking
lap>=0.4.0

# Evaluation
motmetrics>=1.4.0
scipy>=1.10.0

# Configuration
PyYAML>=6.0

# Data Processing
pandas>=2.0.0
